src/requirements.txt
Django==1.8
django-allauth==0.30.0
django-solo==1.1.2
oauthlib==2.0.1
python-openid==2.2.5
requests==2.12.5
requests-oauthlib==0.7.0
